+++ packages/evm/core/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod db;
 mod events;
+pub mod receipt;
 pub mod state_changes;
 pub mod state_commit;
 pub mod state_hash;
